+++ src/modules/helpLists/dto/help-list-create.dto.ts
@@ -6,7 +6,7 @@ export class HelpListCreateDto {
     description: 'List of help request IDs',
     type: 'array',
     items: {
-      type: 'long',
+      type: 'integer',
     },
   })
   helpRequestsIds?: number[];
