+++ src/modules/helpRequests/dto/help-request-create.dto.ts
@@ -5,13 +5,13 @@ import { HelpRequestStatus } from '../help-request-status';
 export class CreateHelpRequestArticleDto {
   @ApiProperty({
     description: 'Article ID received from the article list',
-    type: 'long',
+    type: 'integer',
   })
   readonly articleId!: number;
 
   @ApiProperty({
     description: 'Number of items',
-    type: 'long',
+    type: 'integer',
   })
   readonly articleCount!: number;
 }
